Create basic game resources
This commit creates the index.html and rps.js files to get all the
browser required files in place.

Additionally, rps.js has a "skeleton" of the functions that will
be used at this point to make the game work, based on the
instructions and requirements set in the lesson assignment.