Complete getPlayerChoice() and minor tweaks
Added a prompt for user input in the getPlayerChoice function. This
will now take in the player's selection and make it lower case,
then return the selection.

Also, moved random number expression from the computerChoice variable
to just being in the switch statement expression.

Updated the player selection from a variable to a constant given
that it will not change during execution, after the player has made
their selection.

Finally, updated the return strings in the playRound() function.