Complete game() and minor refactoring
Added logic for the game function. This function plays 5 rounds
and keeps track of the score. It prints the result of each round
and then the winner of the game.

Also changed the return value for the playRound function so that it
is easier to work with the game function and keep track of score.
playRound now returns a simple 'win', 'lose' or 'draw' instead of a
dynamic round win/lose message. The round win/lose message is now
handled by the game function.