Update button logic to make it more compact
I've updated the code to get all player choice buttons as a NodeList
which then adds the click event listener. When the button is clicked,
the play selection is stored in a variable. Additionally, when a
button is clicked, the getComputerChoice function is executed and
the result stored in a variable. The playRound function is then
called with player and computer selections passed as arguments.