Make updates to logic and add helper functions
I added several helper functions to keep track of scores, update
the score board displayed on the HTML page, show a game reset button
after 5 rounds of play and reset the game scores and score board.
Also added the necessary HTML to the page to make it all work and
introduced a styles.css stylesheet to help with rendering UI elements.